Code review feedback: Move Metadata types into separate file and update references
