+++ src/app/pages/MediaArticlePage/MediaArticlePage.tsx
@@ -39,12 +39,12 @@ import RelatedTopics from '../../legacy/containers/RelatedTopics';
 import NielsenAnalytics from '../../legacy/containers/NielsenAnalytics';
 import ArticleMetadata from '../../legacy/containers/ArticleMetadata';
 
+import { OptimoBlock } from '../../models/types/optimo';
 import {
-  OptimoBlock,
   MetadataFormats,
   MetadataTaggings,
   MetadataTopics,
-} from '../../models/types/optimo';
+} from '../../models/types/metadata';
 
 import LinkedData from '../../components/LinkedData';
 import Byline from '../../components/Byline';
